PyTorch/requirements.txt

torch
torchvision
cuda-python
tensorboard
numpy
pillow
future
tqdm
matplotlib
scipy
scikit-learn
gradio